Use @octokit/core instead of octokat
This means one dependency less, and the use of GitHub v3 API will also
get the benfit of @octokit/plugin-throttling.

lib/graphql.js is renamed and generalized to lib/octokit.js. It could
be folded into octokit.js, but is kept separate for ease of testing.

list-unconnected-contributors.js also used octokat, so the needed
functionality is brought into lib/github.js.
