+++ lib/github.js
@@ -2,12 +2,53 @@
 
 "use strict";
 
-const config = require("../config.json");
+const octokit = require("./octokit.js");
 
-const graphql = require("./graphql.js");
-// github API v3 needed to check webhooks
-const Octokat = require("octokat");
-const octo = new Octokat({token: config.ghToken});
+const licenseQuery = `
+  query {
+    repository(owner: "w3c", name: "licenses") {
+      contributing: object(expression: "HEAD:WG-CONTRIBUTING.md") {
+        ... on Blob {
+          text
+        }
+      }
+      contributingSw: object(expression: "HEAD:WG-CONTRIBUTING-SW.md") {
+        ... on Blob {
+          text
+        }
+      }
+      license: object(expression: "HEAD:WG-LICENSE.md") {
+        ... on Blob {
+          text
+        }
+      }
+      licenseSw: object(expression: "HEAD:WG-LICENSE-SW.md") {
+        ... on Blob {
+          text
+        }
+      }
+    }
+  }
+`;
+
+async function w3cLicenses() {
+  let res = await octokit.graphql(licenseQuery);
+  res = res.repository;
+
+  if (res.contributing) {
+    res.contributing = res.contributing.text;
+  }
+  if (res.contributingSw) {
+    res.contributingSw = res.contributingSw.text;
+  }
+  if (res.license) {
+    res.license = res.license.text;
+  }
+  if (res.licenseSw) {
+    res.licenseSw = res.licenseSw.text;
+  }
+  return res;
+}
 
 const repoQuery = `
   query ($org: String!, $cursor: String) {
@@ -102,12 +143,12 @@ const labelQuery = `
 
 async function *listRepos(org) {
   for (let cursor = null; ;) {
-    const res = await graphql(repoQuery, {org, cursor});
+    const res = await octokit.graphql(repoQuery, {org, cursor});
     for (const repo of res.organization.repositories.nodes) {
       const labels = repo.labels.nodes;
       // Fetch more labels if they are paginated
       for (let pageInfo = repo.labels.pageInfo; pageInfo.hasNextPage;) {
-        const res = await graphql(labelQuery, {
+        const res = await octokit.graphql(labelQuery, {
           org,
           repo: repo.name,
           cursor: pageInfo.endCursor
@@ -126,9 +167,23 @@ async function *listRepos(org) {
   }
 }
 
-async function listRepoHooks(org, repo) {
-  const hooks = await octo.repos(`${org}/${repo}`).hooks.fetch();
-  return hooks.items;
+async function listRepoContributors(owner, repo) {
+  const res = await octokit.request('GET /repos/:owner/:repo/contributors', {
+    owner, repo,
+  });
+  return res.data;
 }
 
-module.exports = {listRepos, listRepoHooks};
+async function listRepoHooks(owner, repo) {
+  const res = await octokit.request('GET /repos/:owner/:repo/hooks', {
+    owner, repo,
+  });
+  return res.data;
+}
+
+module.exports = {
+  w3cLicenses,
+  listRepos,
+  listRepoContributors,
+  listRepoHooks,
+};
