+++ lib/octokit.js
@@ -23,14 +23,14 @@ const octokit = new Octokit({
     },
     onAbuseLimit: (retryAfter, options) => {
       if (options.request.retryCount < MAX_RETRIES) {
-        console.warn(`Abuse detected triggered, retrying after ${retryAfter} seconds`)
+        console.warn(`Abuse detection triggered, retrying after ${retryAfter} seconds`)
         return true;
       } else {
-        console.error(`Abuse detected triggered, giving up after ${MAX_RETRIES} retries`);
+        console.error(`Abuse detection triggered, giving up after ${MAX_RETRIES} retries`);
         return false;
       }
     }
   }
 });
 
-module.exports = octokit.graphql;
+module.exports = octokit;
