+++ lib/validator.js
@@ -190,7 +190,7 @@ function validateAshHooks(hooks) {
   // pushed instead of nulls. This is simply to match the original structure
   // of report.json. TODO: change this to null, or add error details.
   const errors = [];
-  const ashHooks = hooks.filter(h => ashnazgHookUrls.includes(h.config.url) && h.config.contentType === "json" && h.config.insecureSsl === "0" && h.config.secret !== "");
+  const ashHooks = hooks.filter(h => ashnazgHookUrls.includes(h.config.url) && h.config.content_type === "json" && h.config.insecure_ssl === "0" && h.config.secret !== "");
   if (ashHooks.length === 0) {
     errors.push(['missingashnazghook', {}]);
   }
