+++ test/github.js
@@ -7,6 +7,59 @@ const proxyquire = require('proxyquire');
 const sinon = require('sinon');
 
 describe('github', () => {
+  describe('w3cLicenses', () => {
+    it('happy path', async () => {
+      const graphql = sinon.fake.resolves({
+        repository: {
+          contributing: {
+            text: 'mock WG-CONTRIBUTING.md content'
+          },
+          contributingSw: {
+            text: 'mock WG-CONTRIBUTING-SW.md content'
+          },
+          license: {
+            text: 'mock WG-LICENSE.md content'
+          },
+          licenseSw: {
+            text: 'mock WG-LICENSE-SW.md content'
+          },
+        }
+      });
+      const github = proxyquire('../lib/github.js', {
+        './octokit.js': {graphql}
+      });
+      const lic = await github.w3cLicenses();
+      assert.deepStrictEqual(lic, {
+        contributing: 'mock WG-CONTRIBUTING.md content',
+        contributingSw: 'mock WG-CONTRIBUTING-SW.md content',
+        license: 'mock WG-LICENSE.md content',
+        licenseSw: 'mock WG-LICENSE-SW.md content',
+      });
+      assert(graphql.calledOnce);
+    });
+
+    it('no files found', async () => {
+      const graphql = sinon.fake.resolves({repository: {}});
+      const github = proxyquire('../lib/github.js', {
+        './octokit.js': {graphql}
+      });
+      const lic = await github.w3cLicenses();
+      assert.deepStrictEqual(lic, {});
+      assert(graphql.calledOnce);
+    });
+
+    it('graphql error', async () => {
+      const graphql = sinon.fake.rejects(new Error('mock error'));
+      const github = proxyquire('../lib/github.js', {
+        './octokit.js': {graphql}
+      });
+      await assert.rejects(github.w3cLicenses(), {
+        message: 'mock error'
+      });
+      assert(graphql.calledOnce);
+    });
+  });
+
   describe('listRepos', () => {
     async function toArray(iterator) {
       const array = [];
@@ -41,7 +94,7 @@ describe('github', () => {
         }
       });
       const github = proxyquire('../lib/github.js', {
-        './graphql.js': graphql
+        './octokit.js': {graphql}
       });
       const res = await toArray(github.listRepos('WICG'));
       assert(graphql.calledOnce);
@@ -99,7 +152,7 @@ describe('github', () => {
         }
       });
       const github = proxyquire('../lib/github.js', {
-        './graphql.js': graphql
+        './octokit.js': {graphql}
       });
       const res = await toArray(github.listRepos('WICG'));
       assert(graphql.calledTwice);
@@ -165,7 +218,7 @@ describe('github', () => {
         }
       });
       const github = proxyquire('../lib/github.js', {
-        './graphql.js': graphql
+        './octokit.js': {graphql}
       });
       const res = await toArray(github.listRepos('WICG'));
       assert(graphql.calledThrice);
@@ -181,22 +234,27 @@ describe('github', () => {
     });
   });
 
-  it('fetchRepoHooks', async () => {
-    class Stubkat {
-      repos() {
-        return {
-          hooks: {
-            async fetch() {
-              return {
-                items: ['mock-hook'],
-              }
-            }
-          }
-        }
-      }
-    }
-    const github = proxyquire('../lib/github.js', {'octokat': Stubkat});
+  it('listRepoContributors', async () => {
+    const request = sinon.stub();
+    request.resolves({data: ['mock-contributor']});
+    const github = proxyquire('../lib/github.js', {'./octokit.js': {request}});
+    const contributors = await github.listRepoContributors('foo', 'bar');
+    assert(request.calledOnceWith('GET /repos/:owner/:repo/contributors', {
+      owner: 'foo',
+      repo: 'bar',
+    }));
+    assert.deepEqual(contributors, ['mock-contributor']);
+  });
+
+  it('listRepoHooks', async () => {
+    const request = sinon.stub();
+    request.resolves({data: ['mock-hook']});
+    const github = proxyquire('../lib/github.js', {'./octokit.js': {request}});
     const hooks = await github.listRepoHooks('foo', 'bar');
+    assert(request.calledOnceWith('GET /repos/:owner/:repo/hooks', {
+      owner: 'foo',
+      repo: 'bar',
+    }));
     assert.deepEqual(hooks, ['mock-hook']);
   });
